Split .h files into .h and .cpp files
