+++ test-rpg/includes/Battle.h
@@ -3,46 +3,17 @@
 
 #include <iostream>
 
+#include "Standard.h"
+#include "Character.h"
+
 namespace rpgText
 {
-    bool CheckBattle(int chance)
-    {
-        int num = (rand() % 100) + 1;
-        if (num <= chance)
-        {
-            return 1;
-        }
-        return 0;
-    }
+    // Checks if Enemy is encountered based on "chance"
+    bool CheckBattle(int chance);
 
-    void StartBattle(const Player &player)
-    {
-        Enemy enemy("Goblin");
-        Print("A ");
-        Print(enemy.name);
-        Print(" is attacking ");
-        Log(player.name);
-        // Log(" is attacking");
-        int attackChoice{};
-        while (enemy.GetStatus())
-        {
-            Print("Enemy's health is ");
-            Log(enemy.health);
-            /* Print("Press 1 to attack or any other to wait: ");
-            std::cin >> attackChoice;*/
-            attackChoice = GetIntWithPrompt("Press 1 to attack or any other to wait: ");
-            if (attackChoice == 1)
-            {
-                enemy.ChangeHealth(-player.attack);
-                Log("You attacked");
-            }
-            else
-            {
-                Log("You waited");
-            }
-        }
-        Log("Enemy Defeated!");
-    }
+    //Initiates battle with enemy
+    void StartBattle(const Player &player);
+
 } // namespace rpgText
 
 #endif
